requirement.txt
biopython==1.79
requests==2.26.0
streamlit==1.5.0
plotly==5.3.1
nglview==2.7.7
